pyproject.toml
[tool.poetry]
name = "abschieben.jetzt"
version = "0.1.0"
description = ""
authors = ["kilimnik <mail@kilimnik.de>"]
license = ""
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
pilmoji = "*"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
server = "src.server:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
